+++ src/command_modules/azure-cli-redis/tests/test_redis_scenario.py
@@ -9,8 +9,8 @@ class RedisCacheTests(ScenarioTest):
     @ResourceGroupPreparer()
     def test_create_redis_cache(self, resource_group):
         name = self.create_random_name(prefix='cli', length=24)
-        self.cmd('az redis create -n {} -g {} -l {} --sku-name {} --sku-family {} --sku-capacity {}'.format(
-            name, resource_group, 'WestUS', 'Basic','C','0'))
+        self.cmd('az redis create -n {} -g {} -l {} --sku {} --vm-size {}'.format(
+            name, resource_group, 'WestUS', 'Basic','C0'))
         self.cmd('az redis show -n {} -g {}'.format(name, resource_group), checks=[
             JMESPathCheck('name', name),
             JMESPathCheck('provisioningState', 'Creating')